Split all yaml to isolated files
